[ofono-binder] Initial implementation. JB#55524
The logic is based on ofono-ril-plugin. Only interfaces up to and
including IRadio@1.2 are fully supported, although bits and pieces
of IRadio@1.4 are already there. Full support for IRadio@1.4 will
be added later as a separate task.

For now, only the standard (as in "defined by Google") IRadio
interfaces are used. Support for vendor-specific extensions is
planned and will be added later.